Pull request 2

+++ ActivityLauncher/src/main/java/com/knziha/polymer/CrashHandler.java
@@ -4,6 +4,8 @@ import android.content.Context;
 import android.content.pm.ApplicationInfo;
 import android.content.pm.PackageInfo;
 import android.content.pm.PackageManager;
+import android.database.sqlite.SQLiteException;
+import android.database.sqlite.SQLiteFullException;
 import android.os.Build;
 import android.util.Log;
 
@@ -26,7 +28,6 @@ import java.util.Locale;
 
 //美滋滋
 public class CrashHandler implements UncaughtExceptionHandler {
-	public static final String TAG = "FatalHandler";
 	/** System default handler */
 	private UncaughtExceptionHandler mDefaultHandler;
 	private static CrashHandler instance;
@@ -94,7 +95,16 @@ public class CrashHandler implements UncaughtExceptionHandler {
 
 	@Override
 	public void uncaughtException(@NonNull Thread thread, @NonNull Throwable exception) {
-		Log.e("ODPlayer","::fatal exception : "+exception.toString());
+		String message = exception.toString();
+		Log.e("PolymPic","::fatal exception : "+ message);
+//		if(message.contains("disk is full")) {
+//			return;
+//		}
+//		if (exception instanceof SQLiteFullException) {
+//			CMN.Log(exception);
+//			AgentApplication.exception = exception;
+//			return;
+//		}
 		Writer writer = new StringWriter();
 		PrintWriter printWriter = new PrintWriter(writer);
 		exception.printStackTrace(printWriter);
@@ -113,16 +123,17 @@ public class CrashHandler implements UncaughtExceptionHandler {
 			try {
 				File log=new File(log_path);
 				File dir = log.getParentFile();
-				dir.mkdirs();
-				if(log.isDirectory()) log.delete();
-				FileOutputStream fos = new FileOutputStream(log_path);
-				fos.write(info_builder.toString().getBytes()); fos.close();
-				new File(dir, "lock").mkdirs();
-			} catch (Exception e) {
-				Log.e(TAG, "an error occured while writing file...", e);
-			}
+				if((dir.isDirectory() || dir.mkdirs()) && dir.getFreeSpace()>1024*1024) {
+					if(log.isDirectory()) log.delete();
+					FileOutputStream fos = new FileOutputStream(log_path);
+					fos.write(info_builder.toString().getBytes()); fos.close();
+					new File(dir, "lock").mkdirs();
+				}
+			} catch (Exception ignored) {  }
 		}
-		CMN.Log("crash catched", GlobalOptions.debug?info_builder.toString():"_"+exception);
+		if(GlobalOptions.debug) {
+			CMN.Log(exception);
+		}
 		postUncaughtException(thread, exception);
 	}
 
@@ -131,7 +142,7 @@ public class CrashHandler implements UncaughtExceptionHandler {
 			if(mDefaultHandler != null)
 				mDefaultHandler.uncaughtException(thread, exception);
 			unRegister();
-		}else {
+		} else {
 			unRegister();
 			System.exit(1);
 		}
